+++ packages/components/fields/creatable-select-field/src/creatable-select-field.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { CreatableSelectField } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const options = [
   { value: 'one', label: 'One' },
