+++ packages/components/fields/creatable-select-field/src/version.js
@@ -1,0 +1,2 @@
+// NOTE: This string will be replaced in the `prepare` script by the `scripts/version.js` file.
+export default '__@UI_KIT_PACKAGE/VERSION_OF_RELEASE__';
